Updates CI workflows and adds mise configuration (#451)
* Updates CI workflows and adds mise configuration

Updates CI workflows to use specific versions of shared workflows, improving stability and reproducibility.

Adds a `mise.toml` file to manage node and pnpm versions, and defines a task for CI post-install to streamline dependency installation.

* Adds build command to Playwright workflow

Adds a build command to the Playwright workflow to ensure the project is built before running tests.

This ensures tests are run against the latest build artifacts.